[POC] [WIP] refactor(files_sharing): move getSuggestions(), getRecommendations() out
I like to reach the goals, but the new state show's how entangled the
code is.

Let's me think about this change - I might park or drop this commit.
Otherwise: let's discuss.

== Goal

* Make suggestion/recommendation code reusable
* Have it as separate module
  * to reduce LOC of SharingInput
  * to make the extracted function easier testable

== Notable changes

Apart from obvious moving code:

1. this.externalResults is actually
   OCA.Sharing.ShareSearch.state.results

   This is now directly referenced.

2. the "lookup" parameter to getSuggestions() was never passed and
   internally defaulted

   The parameter was therefor removed.

== Findings

This reveals a bit of a mess:

1. the filter function depends heavily on the Vue context
2. the formatting of API results for UI display is mixed
   with combining data
3. external results (OCA.Sharing.ShareSearch.state) are
   already formatted

== Refactoring left-overs

1. existingSharesFilter() depends on reshare, linkShares,
   shares from the Vue object

   I'd like to keep this function independant from Vue actually, but
   stop here to not go deeper into a refactoring rabbit hole.

   Marked as TODO.

2. Separation of data fetching and data combining from formatting.
